requirements.txt
aiofiles==23.2.1
alembic==1.12.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.8.1
attrs==25.3.0
Automat==25.4.16
bcrypt==3.2.0
beautifulsoup4==4.12.2
bidict==0.23.1
black==25.1.0
blinker==1.9.0
bottle==0.13.4
Brotli==1.1.0
bs4==0.0.2
cachetools==4.2.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorlog==6.10.1
ConfigArgParse==1.7.1
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
DBUtils==3.0.3
defusedxml==0.7.1
Django==5.2.3
dnspython==2.8.0
docutils==0.21.2
ecdsa==0.19.1
email-validator==2.3.0
ephem==4.2
et_xmlfile==2.0.0
Faker==37.11.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==2.3.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
Kivy==2.3.1
Kivy-examples==3.0.0.dev0
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
locust==2.41.6
locust-cloud==1.27.6
LunarCalendar==0.0.9
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
msgpack==1.1.2
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
narwhals==1.46.0
numpy==2.2.5
opencv-python==4.7.0.72
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parsel==1.10.0
passlib==1.7.4
pathspec==0.12.1
pefile==2023.2.7
platformdirs==4.3.8
playwright==1.55.0
plotly==6.2.0
pluggy==1.6.0
Protego==0.4.0
proxy_tools==0.1.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyMySQL==1.1.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
pystray==0.19.5
pytest-base-url==2.1.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
python-socketio==5.14.2
python-telegram-bot==13.7
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.1.0
queuelib==1.8.0
redis==5.0.1
requests-file==2.1.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
Scrapy==2.13.2
seaborn==0.13.2
selenium==4.34.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
telegram==0.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
tldextract==5.3.0
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.24.0
w3lib==2.3.1
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
zope.event==6.0
zope.interface==7.2
